+++ apps/example/App.tsx
@@ -1,6 +1,6 @@
 import React from 'react';
 import { useFlipperInspector, ReactNativeInspectorOverlay, StoreProvider } from 'react-native-flipper-inspector';
-import SimpleTest from './SimpleTest';
+import NetworkTest from './NetworkTest';
 
 export default function App() {
   // ONE LINE SETUP! 🚀
@@ -8,7 +8,7 @@ export default function App() {
 
   return (
     <StoreProvider>
-      <SimpleTest />
+      <NetworkTest />
       <ReactNativeInspectorOverlay />
     </StoreProvider>
   );